add simple union filter
