+++ src/main/scala/com/kunyandata/backtesting/parser/Rules.scala
@@ -14,14 +14,14 @@ object Rules {
       (1, HashMap(
 
         // 包含大于小于的语句的操作码
-        ("总股本大于x亿", 101),
-        ("总股本小于x亿", 101),
-        ("总股本等于x亿", 101),
-        ("总股本大于x亿小于x亿", 101),
-        ("流通股本大于x亿", 102),
-        ("流通股本小于x亿", 102),
-        ("流通股本等于x亿", 102),
-        ("流通股本大于x亿小于x亿", 102),
+        ("总股本大于x万", 101),
+        ("总股本小于x万", 101),
+        ("总股本等于x万", 101),
+        ("总股本大于x万小于x万", 101),
+        ("流通股本大于x万", 102),
+        ("流通股本小于x万", 102),
+        ("流通股本等于x万", 102),
+        ("流通股本大于x万小于x万", 102),
         ("总市值大于x亿", 103),
         ("总市值小于x亿", 103),
         ("总市值等于x亿", 103),
